Refactor controllers in SimulatorApi
